pyproject.toml
[project]
name = "web-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.3",
    "fake-useragent>=2.2.0",
    "fastapi>=0.124.2",
    "google-api-python-client>=2.187.0",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.14",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
    "youtube-transcript-api>=1.2.3",
]
